AnkiLMS/backend/requirements.txt
django
psycopg2-binary
djangorestframework
django-cors-headers
django-environ
djangorestframework-simplejwt